// src/avatarSvg.gen.ts
export type AvatarVals = Record<(typeof AVATAR_VARS)[number], string | number | boolean>;

export const AVATAR_VARS = ["aBackpack","aBangles","aBeanie","aBindi","aCap","aCrossbody","aGlasses","aHeadphones","aHoops","aJhumkas","aNecklace","aStuds","aSunglasses","aWatch","accent","bAline","bCargo","bJeans","bJoggers","bLeggings","bMidi","bNauvari","bPants","bPencil","bPleated","bShorts","bWide","belt","beltC","beltL","blouseD","borderW","botC","botD","botDenim","botL","capC","capD","full","h","hBangsPart","hBob","hBraid","hCapBack","hCurly","hLong","hPixie","hPony","hSleek","hTie","hair","hairD","hairL","hasOuter","kanji","label","niviBorder","oBadge","oBlazer","oBomber","oCardigan","oDenim","oLeather","oPuffer","oTrench","oVest","outC","outD","outL","palluBack","palluDiag","palluEdge","party","rimFilter","sBoots","sFlats","sHeels","sHigh","sLoafers","sSandals","sSneakers","sheerPallu","sheerSkirt","shoeD","shoeL","shoesOver","shoesUnder","showBun","showBuns","skin","skinD","skinL","skirtSaree","slvC","slvD","slvL","slvLong","slvPuffy","slvShort","slvStripe","sole","tCollar","tHoodie","tSareeB","tSweater","tTank","tTeeBody","tTurtle","topC","topD","topL","uid","w"] as const;